requirements.txt
fastapi[standard]
currencyconverter
motor
python-jose
pytest
pytest-cov
pre-commit
isort
PyJWT
black
pylint
mypy
bandit
pandas-stubs
types-python-jose
matplotlib
pandas
pandas-stubs
reportlab
python-telegram-bot
requests
apscheduler>=3.8.0,<4.0.0
